Support unicode domain names
WARNING: non-schema compliant change.
The default schema for domain and email components are
caseIgnoreIA5SubstringMatch which is strictly IA5 charset.
However, Windchil has inappropriately been storing utf-8 characters
in these fields and so we have no choice but to continue this until
data migration tools and detection tools are created.